explicit cast to avoid warning on gcc 4.6.3

+++ fqd_ccs.c
@@ -158,11 +158,13 @@ fqd_ccs_loop(remote_client *client) {
     if(rv < 0) break;
     t = fq_gethrtime();
     hb_us = ((unsigned long long)client->heartbeat_ms) * 1000000ULL;
-    if(client->heartbeat_ms && client->last_heartbeat < (unsigned long long)(t - hb_us)) {
+    if(client->heartbeat_ms &&
+       (unsigned long long)client->last_heartbeat < (unsigned long long)(t - hb_us)) {
       if(fqd_css_heartbeat(client)) break;
       client->last_heartbeat = t;
     }
-    if(hb_us && client->last_activity < (unsigned long long)(t - hb_us * 3)) {
+    if(hb_us &&
+       (unsigned long long)client->last_activity < (unsigned long long)(t - hb_us * 3)) {
       ERRTOFD(client->fd, "heartbeat failed");
 #ifdef DEBUG
       fq_debug(FQ_DEBUG_CONN, "heartbeat failed from %s\n", client->pretty);
